internet_shop/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.4
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
reportlab==4.2.5
sqlparse==0.5.3
typing_extensions==4.12.2
